feat: update story structure
