+++ packages/date-picker/docs/DatePicker.stories.tsx
@@ -1,15 +1,10 @@
 import React, { useState } from "react"
-import { Story } from "@storybook/react"
-import { usePopper } from "react-popper"
 import { Paragraph } from "@kaizen/typography"
 import { Button } from "@kaizen/button"
 import { FieldMessageStatus } from "@kaizen/draft-form"
 import { CodeBlock } from "@kaizen/design-tokens/docs/DocsComponents"
-import { enAU } from "date-fns/locale"
 import { CATEGORIES, SUB_CATEGORIES } from "../../../storybook/constants"
-import { DatePicker, DayOfWeek, ValidationResponse } from "../src/DatePicker"
-import { Calendar, CalendarProps } from "../src/DatePicker/components/Calendar"
-import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
+import { DatePicker, ValidationResponse } from "../src/DatePicker"
 
 const SUPPORTED_LOCALES = ["en-US", "en-AU"]
 
@@ -168,189 +163,4 @@ export const ValidationStory = props => {
 ValidationStory.storyName = "Validation"
 ValidationStory.parameters = {
   docs: { source: { type: "code" } },
-}
-
-const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
-  isReversed,
-}) => {
-  const [selectedDate, setValueDate] = useState<Date | undefined>()
-
-  return (
-    <>
-      <StoryWrapper isReversed={isReversed}>
-        <StoryWrapper.RowHeader
-          headings={[
-            "Default",
-            "Selected Value",
-            "Custom Description",
-            "Disabled",
-            "Error",
-          ]}
-        />
-        <StoryWrapper.Row rowTitle="Input">
-          <DatePicker
-            id="datepicker-default"
-            labelText="Label"
-            selectedDay={selectedDate}
-            onDayChange={setValueDate}
-            onValidate={() => undefined}
-            isReversed={isReversed}
-            status="default"
-            validationMessage={undefined}
-            locale="en-AU"
-          />
-          <DatePicker
-            id="datepicker-selected"
-            labelText="Label"
-            selectedDay={new Date(2022, 1, 5)}
-            onDayChange={() => undefined}
-            onValidate={() => undefined}
-            isReversed={isReversed}
-            status="default"
-            validationMessage={undefined}
-            locale="en-AU"
-          />
-          <DatePicker
-            id="datepicker-description"
-            labelText="Label"
-            selectedDay={undefined}
-            onDayChange={() => undefined}
-            onValidate={() => undefined}
-            isReversed={isReversed}
-            description={
-              <>
-                <Paragraph
-                  tag="span"
-                  variant="small"
-                  color={isReversed ? "white" : "dark"}
-                >
-                  My <strong>Custom</strong> Description Paragraph
-                </Paragraph>
-              </>
-            }
-            status="default"
-            validationMessage={undefined}
-            locale="en-AU"
-          />
-          <DatePicker
-            id="datepicker-disabled"
-            labelText="Label"
-            selectedDay={undefined}
-            onDayChange={() => undefined}
-            onValidate={() => undefined}
-            isReversed={isReversed}
-            status="default"
-            validationMessage={undefined}
-            locale="en-AU"
-            disabled
-          />
-          <DatePicker
-            id="datepicker-error"
-            labelText="Label"
-            selectedDay={new Date("potato")}
-            onDayChange={() => undefined}
-            onValidate={() => undefined}
-            isReversed={isReversed}
-            status="error"
-            validationMessage="Invalid Date"
-            locale="en-AU"
-          />
-        </StoryWrapper.Row>
-      </StoryWrapper>
-      <StoryWrapper isReversed={isReversed}>
-        <StoryWrapper.RowHeader headings={["en-AU", "en-US"]} />
-        <StoryWrapper.Row rowTitle="Localisation">
-          <DatePicker
-            id="datepicker-enAU"
-            labelText="Label"
-            selectedDay={new Date("2022, 1, 5")}
-            onDayChange={() => undefined}
-            onValidate={() => undefined}
-            isReversed={isReversed}
-            status="default"
-            validationMessage={undefined}
-            locale="en-AU"
-          />
-          <DatePicker
-            id="datepicker-enUS"
-            labelText="Label"
-            selectedDay={new Date("2022, 1, 5")}
-            onDayChange={() => undefined}
-            onValidate={() => undefined}
-            isReversed={isReversed}
-            status="default"
-            validationMessage={undefined}
-            locale="en-US"
-          />
-        </StoryWrapper.Row>
-      </StoryWrapper>
-    </>
-  )
-}
-
-export const StickerSheetDefault = StickerSheetTemplate.bind({})
-StickerSheetDefault.storyName = "Date Input"
-StickerSheetDefault.parameters = {
-  chromatic: { disable: false },
-  controls: { disable: true },
-}
-
-const CalendarExample = (props: Partial<CalendarProps>): JSX.Element => {
-  const [popperElement, setPopperElement] = useState<HTMLDivElement | null>(
-    null
-  )
-
-  const [referenceElement, setReferenceElement] =
-    useState<HTMLDivElement | null>(null)
-
-  const { styles, attributes } = usePopper(referenceElement, popperElement, {
-    modifiers: [
-      {
-        name: "offset",
-        options: {
-          offset: [0, 15],
-        },
-      },
-    ],
-    placement: "bottom-start",
-  })
-
-  return (
-    <div ref={setReferenceElement}>
-      <Calendar
-        mode="single"
-        id="calendar-dialog"
-        setPopperElement={setPopperElement}
-        popperStyles={styles}
-        popperAttributes={attributes}
-        weekStartsOn={DayOfWeek.Sun}
-        onDayChange={() => undefined}
-        defaultMonth={new Date(2022, 1, 5)}
-        locale={enAU}
-        {...props}
-      />
-    </div>
-  )
-}
-
-const StickerSheetCalendarTemplate: Story<{ isReversed: boolean }> = ({
-  isReversed,
-}) => (
-  <StoryWrapper isReversed={isReversed}>
-    <StoryWrapper.RowHeader headings={["Selected Date", "Disabled Dates"]} />
-    <StoryWrapper.Row rowTitle="Calendar">
-      <CalendarExample value={new Date(2022, 1, 5)} />
-      <CalendarExample
-        disabledDays={[new Date(2022, 1, 15), { after: new Date(2022, 1, 17) }]}
-        id="calendar-dialog-disabled"
-      />
-    </StoryWrapper.Row>
-  </StoryWrapper>
-)
-
-export const StickerSheetCalendar = StickerSheetCalendarTemplate.bind({})
-StickerSheetCalendar.storyName = "Calendar"
-StickerSheetCalendar.parameters = {
-  chromatic: { disable: false },
-  controls: { disable: true },
 }
